fixup! lib/vsprintf.c: eliminate potential race in string()

+++ lib/vsprintf.c
@@ -612,7 +612,7 @@ static noinline_for_stack
 char *string(char *buf, char *end, const char *s, struct printf_spec spec)
 {
 	int len = 0;
-	int lim = spec.precision;
+	size_t lim = spec.precision;
 
 	if ((unsigned long)s < PAGE_SIZE)
 		s = "(null)";
